Add tracking for sold listings, bug fixes

+++ src/main/java/com/broadheadenterprises/websitescraper/repository/ItemInfoRepository.java
@@ -3,5 +3,5 @@ package com.broadheadenterprises.websitescraper.repository;
 import com.broadheadenterprises.websitescraper.ItemInfo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemInfoRepository extends CrudRepository<ItemInfo, Long> {
+public interface ItemInfoRepository extends CrudRepository<ItemInfo, Integer> {
 }
